+++ phantomshield-obfuscator/src/main/java/tech/skidonion/obfuscator/dictionary/impl/SpacesDictionary.java
@@ -30,11 +30,11 @@ public class SpacesDictionary extends Dictionary {
     }
 
     @Override
-    public String generate(int index) {
+    public String generate(long index) {
         int totalCharacterCount = size();
 
-        int baseIndex = index / totalCharacterCount;
-        int offset = index % totalCharacterCount;
+        int baseIndex = (int) (index / totalCharacterCount);
+        int offset = (int) (index % totalCharacterCount);
 
         char newChar = CHARSET[offset];
